Include a first cut of a Periodic handler for running tasks in the client
There are a number of tasks which we're going to run periodically and this
simple class will act as a consistent interface to make sure that everything is
being executed properly.

Periodic uses an instance-specific minute offset for hourly tasks to ensure that
all clients are not running periodic tasks, such as sending data to the server
at the top of the hour, and so on.