[JENKINS-52260] Try to push the update level a few times before abandoning
As we can see pretty regularly in CI, the client would fail to start
because the server was not able to serve an update.
This was because of a race condition: the backend could still be starting
up when the `curl` call occurs.

This change retries this call a few times with some backoff strategy.

This should avoid seeing in the build logs:
>
`curl: (56) Recv failure: Connection reset by peer`

and then the client failing to update, since the backend has *no*
update level to serve:

```
error: Failed to query updates { Error: Cannot read property 'docker-cloud' of undefined
    at convert (/evergreen/client/node_modules/@feathersjs/errors/lib/index.js:240:7)
    at toError (/evergreen/client/node_modules/@feathersjs/rest-client/lib/base.js:12:9)
    at <anonymous>
    at process._tickCallback (internal/process/next_tick.js:188:7)
  status: 'ERROR',
  response:
   Response {
     size: 0,
     timeout: 0,
     [Symbol(Body internals)]: { body: [Object], disturbed: true, error: null },
     [Symbol(Response internals)]:
      { url: 'http://backend:3030/update/50b2fef8-63ee-442b-9910-e2c67f693b08?level=0',
        status: 500,
        statusText: 'Internal Server Error',
        headers: [Object] } } } undefined undefined
2018-07-17 18:25:42,610 INFO success: evergreen-client entered RUNNING state, process has stayed up for > than 2 seconds (startsecs)
```