Pass the artifactId back in the Update API response so the client can use it for a file name
The previous behavior was to rely on the URL for the artifact name which would
result in artifacts coming down with versioned names, e.g. essentials-3.0.hpi
and then dropped on the filesystem.

Of course, this would cause epic problems when an update occurred as the
JENKINS_HOME would contain essentials-3.0.hpi and essentials-3.1.hpi.

With this change, the artifacts will simply be saved as `${artifactId}.hpi`