Force the reconnect on the evergreen-client socket.io connection
From my read on the socket.io 2 documentation, the reconnection to infinity
should be the default, but in practice that does not appear to be the case.

Tested this locally by killing the server for about ten minutes and then
bringing it back. Once the server was back online, the client reconnects and
receives the pings appropriately

Should fix JENKINS-53057