Don't intentionally override already computed plugins to pass back to the client
In the case where the client, with a flavor, had published an empty versions
manifest, the filterVersionsForClient function would override the already
computed plugins needed for that flavor with the base set of plugins.

This avoids that behavior by making the function no-op when there are no plugins
in the versions manifest.