Utilize the dependency graph to remove non-optional plugins which haven't been "upgraded"
If a plugin has an optional dependency which specifies a later version than some
other  non-optional dependency, we want to take the later of the two.

However, we do not want to include dependencies from optional dependencies which
are not otherwise referenced in the graph.

With this change the `needed` list contains all the latest versions of software
from the graph which we may need, and then by flatting and filtering against the
tree we can determine which plugins are actually required from the fully
realized dependency graph.