[JENKINS-53215] Configure Secure defaults OOTB: test only
Jenkins Evergreen should have no Admin monitor showing under /manage after
startup.

Note:
for now, this does not fix the
`Jenkins root URL is empty but is required for the proper operation of many Jenkins features like email notifications, PR status update, and environment variables such as BUILD_URL.`
monitor.
This one has been fixed only very recently in the core itself, and
requires us to guide users to the /configure page and save it, since
we cannot know in advance what the value is going to be, at least for the
Docker Cloud flavor.
We might be able to do it though for the AWS flavor.
That is not a security related issue however, so it's not very critical
at this point.