Relocate supervisord to the flavored images and make base much more explicit
The base image is not really sensible for end-users in any way and only serves
as a means of consolidating system dependencies before the evergreen-client and
remaining distribution is installed.

This change makes this relationship much more explicit than it was previously,
since the backend would reject flavor-less registrations anyways.

Once this change is merged, jenkins/evergreen:latest should be deleted from
Dockerhub