This time fix the update service behavior when considering a flavor for real
My previous attempt at this fix was unsuccessful because I fabricated the
environment in the unit tests incorrectly. The underlying problem is that the
filterVersionsForClient() method is expecting that the `computedManifest` has
already had the flavor'ed plugins added to the manifest.

The `record` is completely irrelevant, and doesn't have the flavor changes
appropriately applied

Fixes JENKINS-53318