Stop using a volume for the client
Because of this, our tests are taking a very high risk to run against
something that will actually *not* be the final image/container content.

We must remove this, and possibly add an Docker Compose override file
for use with `make run`, but certainly stop using this volume mount for
automated testing.