*Now* the test does test what I wanted it to...
Basically, I wrote the successive then and catch like every steps were
going to be gone through. But obviously, this is not how Promises work:
if a 304 is thrown at the first line, it was going to the last catch
directly, making the test succeed, but absolutely not testing the
intended use case.