Add workaround for the race condition
If the borked UL upload happens *before* the of healthchecking for the
first startup, then test will fail because the new UL will be skipped.
That sleep shifts the new UL upload enough to make it unlikely to happen.

Real durable fix will be to scan logs for the end and success of the
healthcheck, not just the usual "Jenkins is fully up and running" log