Rework directory absence check
On alpine or not, when a dir is missing, the message is not exactly the
same. So this change does not check anymore the message, but relies on
`test -d` failure instead.

Here is a more concrete example of the difference:

```
$ docker run node:10 ls /home/non-existing
ls: cannot access /home/non-existing: No such file or directory
$ docker run node:10-alpine ls /home/non-existing
ls: /home/non-existing: No such file or directory
```